Fixed linter issues

+++ actions/delegate_user_stake.go
@@ -84,7 +84,7 @@ func (s *DelegateUserStake) Execute(
 	}
 
 	// Get current time
-	currentTime := time.Unix(int64(timestamp), 0).UTC()
+	currentTime := time.Unix(timestamp, 0).UTC()
 	// Convert Unix timestamps to Go's time.Time for easier manipulation
 	startTime := time.Unix(int64(s.StakeStartTime), 0).UTC()
 	if startTime.Before(currentTime) {
